Split files so changes are better tracked
